refactor: refactored reviews screen.
