--- ui/src/main/java/com/cairosquad/ui/details/ReviewsScreen.kt
@@ -1,186 +1,0 @@
-package com.cairosquad.ui.details
-
-import android.os.Build
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.WindowInsets
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.statusBars
-import androidx.compose.foundation.layout.systemBarsPadding
-import androidx.compose.foundation.layout.windowInsetsPadding
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.collectAsState
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.BlurredEdgeTreatment
-import androidx.compose.ui.draw.blur
-import androidx.compose.ui.graphics.vector.ImageVector
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.res.vectorResource
-import androidx.compose.ui.unit.dp
-import androidx.hilt.navigation.compose.hiltViewModel
-import com.cairosquad.design_system.basic_component.AppBar
-import com.cairosquad.design_system.basic_component.SnackBar
-import com.cairosquad.design_system.modifier.dropShadow
-import com.cairosquad.design_system.theme.Theme
-import com.cairosquad.ui.R
-import com.cairosquad.ui.movio_component.LoadingReviewCard
-import com.cairosquad.ui.movio_component.ReviewCard
-import com.cairosquad.ui.navigation.LocalNavController
-import com.cairosquad.ui.utils.ObserveAsEffect
-import com.cairosquad.viewmodel.details.reviews.ReviewsEffect
-import com.cairosquad.viewmodel.details.reviews.ReviewsInteractionListener
-import com.cairosquad.viewmodel.details.reviews.ReviewsScreenState
-import com.cairosquad.viewmodel.details.reviews.ReviewsViewModel
-import kotlinx.coroutines.Dispatchers
-
-@Composable
-fun ReviewsScreen(
-    mediaId: Long,
-    isMovie: Boolean,
-) {
-	val viewModel: ReviewsViewModel =
-        hiltViewModel<ReviewsViewModel, ReviewsViewModel.Factory> { factory ->
-            factory.create(
-                mediaId = mediaId,
-                isMovie = isMovie,
-                dispatcher = Dispatchers.IO
-            )
-        }
-    val navController = LocalNavController.current
-    val state = viewModel.screenState.collectAsState()
-
-	LaunchedEffect(Unit) {
-		viewModel.getReviews()
-	}
-
-	ObserveAsEffect(viewModel.effect) { effect ->
-		when (effect) {
-			is ReviewsEffect.NavigateBack -> navController.popBackStack()
-		}
-
-	}
-
-	ReviewsContent(
-		listener = viewModel,
-		state = state.value,
-	)
-
-	if (state.value.error != null) {
-		SnackBar(
-			imageVector = ImageVector.vectorResource(com.cairosquad.design_system.R.drawable.danger),
-			message = state.value.error ?: stringResource(R.string.something_went_wrong),
-			action = {}
-		)
-	}
-}
-
-@Composable
-private fun ReviewsContent(
-	listener: ReviewsInteractionListener,
-	state: ReviewsScreenState
-) {
-	when {
-		state.isLoading -> {
-			ReviewsLoadingContent()
-		}
-
-		state.error != null -> {
-			DetailsFailContent(onTryAgainClick = listener::onRefresh)
-		}
-
-		else -> Box {
-			Box(
-				modifier = Modifier
-					.windowInsetsPadding(WindowInsets.statusBars)
-					.size(230.dp)
-					.align(Alignment.TopEnd)
-					.then(
-						if (Build.VERSION.SDK_INT < Build.VERSION_CODES.S) {
-							Modifier.dropShadow(
-								shape = CircleShape,
-								color = Theme.color.surfaces.onSurfaceAt5,
-								blur = 264.dp,
-								offsetX = 0.dp,
-								offsetY = 0.dp,
-								alpha = 0.10f
-							)
-						} else {
-							Modifier
-								.blur(264.dp, edgeTreatment = BlurredEdgeTreatment.Unbounded)
-								.background(
-									color = Theme.color.surfaces.onSurfaceAt5,
-									shape = CircleShape
-								)
-						}
-					)
-			)
-			Column(
-				modifier = Modifier
-					.fillMaxSize()
-
-					.systemBarsPadding(),
-				horizontalAlignment = Alignment.CenterHorizontally
-			) {
-				AppBar(
-					title = stringResource(R.string.reviews),
-					onBackButtonClicked = listener::onClickBack,
-				)
-				LazyColumn(
-					modifier = Modifier.padding(
-						top = 12.dp,
-						bottom = 12.dp,
-						start = 16.dp,
-						end = 16.dp
-					),
-					verticalArrangement = Arrangement.spacedBy(12.dp)
-				) {
-					items(state.reviews) { review ->
-						ReviewCard(
-							modifier = Modifier.fillMaxWidth(),
-							imgUrl = review.reviewerImageUrl,
-							rating = review.rating,
-							reviewDate = review.reviewDate,
-							reviewText = review.reviewText,
-							reviewerName = review.reviewerName
-						)
-					}
-				}
-			}
-		}
-
-	}
-}
-
-@Composable
-private fun ReviewsLoadingContent() {
-	Column(
-		modifier = Modifier
-			.fillMaxWidth()
-			.systemBarsPadding(),
-		horizontalAlignment = Alignment.CenterHorizontally
-	) {
-		LazyColumn(
-			modifier = Modifier.padding(
-				top = 12.dp,
-				bottom = 12.dp,
-				start = 16.dp,
-				end = 16.dp
-			),
-			verticalArrangement = Arrangement.spacedBy(12.dp)
-		) {
-			items(12) {
-				LoadingReviewCard()
-			}
-		}
-	}
-}
