+++ ui/src/main/java/com/cairosquad/ui/details/reviews/content/ReviewsLoadingContent.kt
@@ -1,0 +1,37 @@
+package com.cairosquad.ui.details.reviews.content
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.systemBarsPadding
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+import com.cairosquad.ui.movio_component.LoadingReviewCard
+
+@Composable
+fun ReviewsLoadingContent() {
+    Column(
+        modifier = Modifier
+            .fillMaxWidth()
+            .systemBarsPadding(),
+        horizontalAlignment = Alignment.CenterHorizontally
+    ) {
+        LazyColumn(
+            modifier = Modifier.padding(
+                top = 12.dp,
+                bottom = 12.dp,
+                start = 16.dp,
+                end = 16.dp
+            ),
+            verticalArrangement = Arrangement.spacedBy(12.dp)
+        ) {
+            items(12) {
+                LoadingReviewCard()
+            }
+        }
+    }
+}
