+++ ui/src/main/java/com/cairosquad/ui/details/reviews/content/ReviewsScreenContent.kt
@@ -1,0 +1,28 @@
+package com.cairosquad.ui.details.reviews.content
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.runtime.Composable
+import com.cairosquad.ui.details.DetailsFailContent
+import com.cairosquad.viewmodel.details.reviews.ReviewsInteractionListener
+import com.cairosquad.viewmodel.details.reviews.ReviewsScreenState
+
+@Composable
+fun ReviewsContent(
+    listener: ReviewsInteractionListener,
+    state: ReviewsScreenState
+) {
+    when {
+        state.isLoading -> {
+            ReviewsLoadingContent()
+        }
+
+        state.error != null -> {
+            DetailsFailContent(onTryAgainClick = listener::onRefresh)
+        }
+
+        else -> Box {
+            BlurredCircle()
+            Reviews(listener, state)
+        }
+    }
+}
